requirements.txt
openai
langchain
faiss-cpu
tiktoken
PyPDF2
streamlit
python-dotenv
